Movie Search App
With the data provided from a movie dataset, research can be provided for the desired movie. If the name of the selected or searched movie is written and pressed on it, details are given. These details consist of a few content such as year, actor, subject. Javascript, HTML, CSS are used in this area.